+++ src/Classes/SDOSFLEX.h
@@ -1,8 +1,0 @@
-//
-//  SDOSFLEX.h
-//
-//  Copyright © 2018 SDOS. All rights reserved.
-//
-
-#import "UIApplication+SDOSFLEX.h"
-