+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [1.3.0 Support Swift Package Manager](https://github.com/SDOSLabs/SDOSFLEX/tree/1.3.0)
+
+- Add support to Swift Package Manager
+
 ## [1.2.0 Update Flex version](https://github.com/SDOSLabs/SDOSFLEX/tree/v1.2.0)
 
 - Updated podspec to have FLEX 4 as dependency